Create the necessary files and directories
The general file structure was created. The index file for the homepage,
as well as an images folder, and a recipes folder with 3 html files for
the 3 recipes that will be added later in the project. Images were also
included in the respective folder, 1 image for each recipe, and an image
to be added to the homepage.